app/requirements.txt
django==3.1.1
gunicorn==20.0.4
djangorestframework==3.12.4
python-decouple==3.4
psycopg2-binary==2.9.9
django-jazzmin
shortuuid